+++ src/Page.php
@@ -87,9 +87,9 @@ class Page {
             $tab_obj->set_name( $tab );
         }
 
-        // if position is already used, add + 1.
-        if( isset( $this->tabs[$position]) ) {
-            $position++;
+        // if position is already used, search for the next free index.
+        if( isset( $this->tabs[$position] ) ) {
+            $position = Helper::get_next_free_index_in_array( $this->tabs, $position );
         }
 
         // add the tab to the list of tabs of these settings.